[#36] refactor: 영수증 페이지 공통 컴포넌트 적용 및 컴포넌트 분리
