+++ src/components/receipt/FailureCard.jsx
